move `FindPathConfig` to `find_path.rs`
